Use CSS generated from bundler
