+++ js/App.svelte
@@ -1,4 +1,5 @@
 <script>
+    import { onMount } from 'svelte';
     import HeartBeat from './HeartBeat.svelte';
     import SidePanel from './SidePanel.svelte';
     import DebugButton from './DebugButton.svelte';
@@ -18,12 +19,64 @@
     let heartbeatOpacity = 0;
 
     let mousePos = null;
+    let moving = false;
+    let movingItem = false;
+    let buildingConveyor = null;
+    let dragStart = null;
+    let dragLast = null;
     let canvas;
     let time = 0;
 
     if(serverSync && !sessionId){
-        getSessionId(id => sessionId = id);
+        getSessionId({
+            port,
+            baseUrl,
+            game,
+            setSessionId: id => sessionId = id,
+        });
     }
+
+    function toLogicalCoords(clientX, clientY) {
+        const r = canvas.getBoundingClientRect();
+        const x = clientX - r.left;
+        const y = clientY - r.top;
+        return [x, y];
+    }
+
+    function pointerMove(evt) {
+        const [x, y] = mousePos = toLogicalCoords(evt.clientX, evt.clientY);
+        game.set_cursor(x, y);
+        infoResult = game.get_info(x, y);
+        if (buildingConveyor) {
+            const elem = document.getElementById("conveyor");
+            if (elem?.checked) {
+                try {
+                    game.preview_build_conveyor(buildingConveyor[0], buildingConveyor[1], x, y, true);
+                }
+                catch (e) {
+                    console.error(`build_conveyor: ${e}`);
+                }
+            }
+        }
+        if (dragStart) {
+            if (dragLast) {
+                game.pan(x - dragLast[0], y - dragLast[1]);
+                dragLast = [x, y];
+            }
+            else {
+                const dragDX = Math.abs(x - dragStart[0]);
+                const dragDY = Math.abs(y - dragStart[1]);
+                // Determine mouse drag (or panning with a touch panel) or a click (or a tap) by checking moved distance
+                if (10 < Math.max(dragDX, dragDY)) {
+                    dragLast = dragStart;
+                }
+            }
+        }
+    }
+
+    onMount(() => {
+        canvas.addEventListener('pointermove', pointerMove);
+    });
 
     setInterval(() => {
         // Increment time before any await. Otherwise, this async function runs 2-4 times every tick for some reason.
@@ -39,6 +92,7 @@
         game.render(ctx);
         if (mousePos !== null) {
             const info = game.get_info(mousePos[0], mousePos[1]);
+            console.log(`${info}`);
             infoResult = info;
         }
         if (websocket) {
